Add support for injecting of definition storage to ezcWorkflowDatabaseExecution
When we need support for prefixed tables, in current implementation it is not possible
to specify options which should be passed to the constructor of ezcWorkflowDatabaseDefinitionStorage
instantiated by default in constructor of ezcWorkflowDatabaseExecution. If definition storage
is runtime requirement of the class ezcWorkflowDatabaseExecution, it should be possible
to inject our own implementation.